% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaper.R
\name{heatmaper}
\alias{heatmaper}
\title{Hetmap function
Create a heatmap based on data with specified cuffoff values}
\usage{
heatmaper(data, filter = "PValue", cuttoff = 0.05, start = 10,
  y.text = ggplot2::element_blank())
}
\arguments{
\item{data}{Dataframe with expression data and HITS-CLIP data}

\item{filter}{Filter based on PValue or FDR}

\item{cuttoff}{Cuttoff value to use with filter}

\item{y.text}{Labels for gene names}
}
\description{
Hetmap function
Create a heatmap based on data with specified cuffoff values
}
