% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEapp.R
\name{runDEAHCapp}
\alias{runDEAHCapp}
\title{A function to run the DE_AHC shiny app}
\usage{
runDEAHCapp(...)
}
\arguments{
\item{...}{Any argument that you can pass to shiny::runApp}
}
\value{
Plot

Runs the shiny app.
}
\description{
This function runs the app
}
\examples{
# microRNA.TEvMP()
}
